+++ docs/source/changelog.rst
@@ -3,6 +3,37 @@
 Release Notes
 #############
 
+1.4.0
+=====
+
+:Release Date: 2025-06-18
+
+* Adds a method for miscellaneous settings to the UHF readers
+* Update minimum reader firmware versions
+
+  * DeskID NFC 2.4
+  * DeskID UHF ETSI 1.5
+  * DeskID UHF FCC 1.6
+  * PLRM 1.13
+  * PulsarLR 1.6
+  * QR NFC 1.3
+  * QRG2 ETSI 1.10
+  * QRG2 FCC 1.9
+  * DWARFG2 V2 1.4
+  * DWARFG2 MINI V2 1.5
+  * DwarfG2 XR v2 1.4
+
+* PulsarLR: Fixed connecting via hostname
+* UHF Gen 2: set_inventory_settings() now supports configuring the
+  RSSI threshold.
+* Support the reset() method both for ASCII and AT readers
+* UhfTag.get_inventory_epc() returns the EPC originally reported
+  by the inventory, which may differ from the EPC reported by
+  a read operation.
+* Python 3.9 is the new minimum supported version
+* Fixed repeated calls to UhfReaderAT.set_inventory_settings().
+  If not all settings were passed every time the method could fail.
+
 1.3.4
 =====
 
@@ -24,6 +55,7 @@ Release Notes
 
 * Fixed read and write requests on legacy readers
 * Added new NFC reader features
+
   * HID keyboard mode
   * NDEF read/write functions
 
@@ -43,10 +75,12 @@ Release Notes
 * Updated class and method descriptions
 * Updated code examples
 * Added new UHF reader features
+
   * Select tag function
   * Short range and protected mode
   * Channel mask settting
   * Device temperature query
+
 * Various bugfixes and improved error handling
 
 1.2.0
@@ -59,6 +93,7 @@ Release Notes
 * tag getter and setter methods updated
 * uhf ascii reader - inventory bug fixed
 * uhf at reader
+
   * phase information now available
   * mask commands updated